### 2024-08-26 #7142 @editorialbot
[PRE REVIEW]: lpcde: Estimation and Inference for Local Polynomial Conditional Density Estimators
**Submitting author:** @rajitachandak (<a href="http://orcid.org/0009-0006-4289-2520">Rajita Chandak</a>)
**Repository:** https://github.com/nppackages/lpcde
**Branch with paper.md** (empty if default branch):
**Version:** v.0.1.4
