+++ R/lpcde/DESCRIPTION
@@ -9,7 +9,7 @@ Authors@R: person(given = "Rajita",
 Maintainer: Rajita Chandak <rchandak@princeton.edu>
 Description: Tools for estimation and inference of conditional densities, derivatives and functions. This is the companion software for Cattaneo, Chandak, Jansson and Ma (2024) <doi:10.48550/arXiv.2204.10359>.
 Depends: R (>= 3.3.0)
-License: GPL-2
+License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: false
 SystemRequirements: GNU make
