### 2024-09-10 #7142 @spholmes
Hi @mayer79,
would you be interested in reviewing this submission to JOSS?
Many Thanks,
Susan
