### 2024-09-10 #7142 @spholmes
Hi @nhejazi,
We have a submission of an R package and article to JOSS, maybe you could help us review it?
Many Thanks
Susan
