### 2024-09-11 #7142 @nhejazi
hi @spholmes -- this looks interesting and i'd be happy to help but i'm bandwidth-limited at the moment. would it be alright if i co-review with a doctoral student i work with (@salbalkus)? he's been studying conditional density estimation in causal inference
