### 2024-09-17 #7241 @editorialbot
[REVIEW]: lpcde: Estimation and Inference for Local Polynomial Conditional Density Estimators
**Submitting author:** @rajitachandak (<a href="http://orcid.org/0009-0006-4289-2520">Rajita Chandak</a>)
**Repository:** https://github.com/nppackages/lpcde
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.6
