### 2024-09-17 #7142 @spholmes
@salbalkus : the review will be on a different issue thread: https://github.com/openjournals/joss-reviews/issues/7241
you can generate your checklist by asking
`@editorialbot generate my checklist` over there.
