### 2024-10-03 #7241 @salbalkus
Hi @rajitachandak, I have a few substantive comments on the software that I'd like to post as issues in the GitHub repo, as recommended by JOSS. Would you mind enabling the "Issues" tab on your package's repo so that I can do so? Thanks!
