### 2024-10-04 nppackages/lpcde#5 @salbalkus
JOSS review: Add community guidelines
Raising as part of JOSS review openjournals/joss-reviews#7241

While there is a contact email, I could not find any information or guidelines for those looking to *contribute* to the software anywhere -- JOSS requires them. I would recommend explaining in the top-level READ.md to inform visitors to the GitHub repo how they can contribute, report issues, or seek support.
