### 2024-10-04 nppackages/lpcde#6 @salbalkus
JOSS review: Fix license
Raising as part of JOSS review openjournals/joss-reviews#7241

Though the MIT license is contained in R/lpcde/LICENSE.md, it looks like the LICENSE plain text file does not contain a valid license, which is required by JOSS. In addition to updating the plain text file, I would recommend moving or copying the license to the top level of the repo as well, so that it is clear the license applies to all code contained within it and is easy to find. I think this is why it was not detected automatically by the editorial bot.
