### 2024-10-04 nppackages/lpcde#7 @salbalkus
JOSS review: Software paper citations and spelling
Raising as part of JOSS review openjournals/joss-reviews#7241

I found a few small issues in the paper:
1. Citations are missing DOI numbers
2. Title of "Cattaneo, M. D., Chandak, R., Jansson, M., & Ma, X. (2024a). Local polynomial conditional density estimators. Bernoulli, 30(4), 3193–3223" I believe should be corrected to say "**Boundary adaptive** local polynomial conditional density estimators"
3. Spelling mistake: "conditioal PDF" at the top of page 2
