### 2024-10-04 nppackages/lpcde#8 @salbalkus
JOSS review: Unit tests
Raising as part of JOSS review openjournals/joss-reviews#7241

Good job including automated testing -- I was able to run the tests via the REPL on my machine. Two notes regarding the unit tests:

1. When I ran the tests, several warnings were printed. These included "Using `size` aesthetic for lines was deprecated in ggplot2 3.4.0. Please use `linewidth` instead." as well as several warnings that objects are longer than expected. These are probably worth amending just to ensure the software is totally robust going into the future.
2. Though you do run the `lpcde` procedure on some small toy examples, there appear to be no unit tests screening the basic statistical properties of the conditional density estimators (bias, variance, etc.) to ensure their output is correct. While I understand that the estimator can be somewhat computationally intensive to run over repeated simulations to check something like empirical variance matching the estimated variance, I would recommend at least checking a few basic properties (predicted densities are between 0 and 1, bias is not too extreme, etc.) on a small simulation with a set seed.
