### 2024-10-04 nppackages/lpcde#7 @salbalkus
Oh, and one more comment: I would also recommend citing the R package / JOSS paper [haldensify: Highly adaptive lasso conditional density estimation in R](https://joss.theoj.org/papers/10.21105/joss.04522) by Hejazi, van der Laan, and Benkeser (2022) and possibly adding it to your table of comparisons.
