### 2024-10-04 nppackages/lpcde#10 @salbalkus
On a somewhat related note, there's a slight typo in the `lpbwcde` documentation: "conditionaldensity" should be "conditional density"
