### 2024-10-04 #7241 @salbalkus
Hi @rajitachandak my first pass is now complete; overall, the paper and the software look great. The remaining unchecked items pertain to some issues that I've filed above; @spholmes I assume I should wait to check these until the issues have been addressed? Thanks!
