+++ paper.bib
@@ -4,8 +4,9 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 @article{CCJM_2024_Bernoulli,
-    title={Local Polynomial Conditional Density Estimators},
+    title={Boundary adaptive local polynomial conditional density estimators},
     author={Cattaneo, Matias D. and Chandak, Rajita and Jansson, Michael and Ma, Xinwei},
+    doi={10.3150/23-BEJ1711},
     journal={Bernoulli},
     volume={30}, number={4}, pages={3193-3223}, year={2024}
 }
@@ -26,6 +27,7 @@
   pages =         {767--779},
   title =         {On the Effect of Bias Estimation on Coverage Accuracy
                    in Nonparametric Inference},
+                   doi = {10.1080/01621459.2017.1285776},
   volume =        {113},
   year =          {2018},
 }
@@ -38,6 +40,7 @@
   pages =         {2998--3022},
   title =         {Coverage Error Optimal Confidence Intervals for Local
                    Polynomial Regression},
+                   doi = {10.3150/21-bej1445},
   volume =        {28},
   year =          {2022},
 }
@@ -45,36 +48,42 @@
 @article{DeGooijer-Zerom_2003_SN,
   title={On Conditional Density Estimation},
   author={De Gooijer, Jan G and Zerom, Dawit},
+  doi={10.1111/1467-9574.00226},
   journal={Statistica Neerlandica}, volume={57}, number={2}, pages={159--176}, year={2003}
 }
 
 @book{Fan-Gijbels_1996_Book,
     title	={Local Polynomial Modelling and Its Applications},
     author	={Fan, Jianqing and Gijbels, Irene},
+    doi={10.1201/9780203748725},
     publisher={Chapman \& Hall/CRC}, year	={1996}
 }
 
 @article{Fan-Yao-Tong_1996_Biometrika,
     title={Estimation of Conditional Densities and Sensitivity Measures in Nonlinear Dynamical Systems},
     author={Fan, Jianqing and Yao, Qiwei and Tong, Howell},
+    doi={10.1093/biomet/83.1.189},
     journal={Biometrika}, volume={83}, number={1}, pages={189--206}, year={1996}
 }
 
 @article{Hall-Racine-Li_2004_JASA,
     title={Cross-Validation and the Estimation of Conditional Probability Densities},
     author={Hall, Peter and Racine, Jeff and Li, Qi},
+    doi={10.1198/016214504000000548},
     journal={Journal of the American Statistical Association}, volume={99}, number={468}, pages={1015--1026}, year={2004}
 }
 
 @article{Hall-Wolff-Yao_1999_JASA,
     title={Methods for Estimating a Conditional Distribution Function},
     author={Hall, Peter and Wolff, Rodney CL and Yao, Qiwei},
+    doi={10.2307/2669691},
     journal={Journal of the American Statistical Association}, volume={94}, number={445}, pages={154--163}, year={1999}
 }
 
 @book{scott2015multivariate,
   title={Multivariate Density Estimation: Theory, Practice, and Visualization},
   author={Scott, David W},
+  doi={10.2307/1270280},
   year={2015},
   publisher={John Wiley \& Sons}
 }
@@ -82,6 +91,7 @@
 @book{simonoff2012smoothing,
   title={Smoothing Methods in Statistics},
   author={Simonoff, Jeffrey S},
+  doi={10.1007/978-1-4612-4026-6},
   year={2012},
   publisher={Springer--Verlag}
 }
[…]
