+++ R/lpcde/tests/testthat/test-lpcde.R
@@ -1,7 +1,3 @@
-test_that("error checking", {
-
-})
-
 test_that("lpcde default output", {
   set.seed(42)
   n=100
