+++ R/lpcde/DESCRIPTION
@@ -14,7 +14,7 @@ Encoding: UTF-8
 LazyData: false
 SystemRequirements: GNU make
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.1.2
+RoxygenNote: 7.3.2
 Imports:
     Rcpp (>= 0.12.8),
     ggplot2,
