### 2024-10-18 d0fb624
working on cov_flag option
