+++ R/lpcde/R/lpcde.R
@@ -149,6 +149,7 @@ lpcde = function(x_data, y_data, y_grid=NULL, x=NULL, bw=NULL, p=NULL, q=NULL,
     }
   }
 
+
   #sd_y = stats::sd(y_data)
   #sd_x = apply(x_data, 2, stats::sd)
   #mx = apply(x_data, 2, mean)
