+++ R/lpcde/R/lpcde_fns.R
@@ -49,18 +49,16 @@ lpcde_fn = function(y_data, x_data, y_grid, x, p, q, p_RBC, q_RBC, bw, mu, nu,
   est_flag = f_hat_val$singular_flag
 
   # standard errors
-  if(cov_flag=="full"){
-  covmat = cov_hat(x_data=x_data, y_data=y_data, x=x, y_grid=y_grid, p=p, q=q,
-                   mu=mu, nu=nu, h=bw, kernel_type=kernel_type)
-  covMat = covmat$cov
-  c_flag = covmat$singular_flag
-  se = sqrt(abs(diag(covMat)))*sd_y*mean(sd_x)
-  } else if (cov_flag == "diag"){
-
-  } else if (cov_flag == "off"){
+  if(cov_flag == "off"){
     covMat = NA
     c_flag = FALSE
     se = rep(NA, length(y_grid))
+  } else {
+    covmat = cov_hat(x_data=x_data, y_data=y_data, x=x, y_grid=y_grid, p=p, q=q,
+                     mu=mu, nu=nu, h=bw, kernel_type=kernel_type, cov_flag=cov_flag)
+    covMat = covmat$cov
+    c_flag = covmat$singular_flag
+    se = sqrt(abs(diag(covMat)))*sd_y*mean(sd_x)
   }
 
   if (rbc){
@@ -83,20 +81,17 @@ lpcde_fn = function(y_data, x_data, y_grid, x, p, q, p_RBC, q_RBC, bw, mu, nu,
       # covariance matrix
 
       # standard errors
-      if(cov_flag=="full"){
-      covmat_rbc = cov_hat(x_data=x_data, y_data=y_data, x=x, y_grid=y_grid,
-                       p=p_RBC, q=q_RBC, mu=mu, nu=nu, h=bw,
-                       kernel_type=kernel_type)
-      covMat_rbc = covmat_rbc$cov
-      c_rbc_flag = covmat_rbc$singular_flag
-      se_rbc = sqrt(abs(diag(covMat_rbc)))*sd_y*mean(sd_x)
-      } else if (cov_flag == "diag"){
-
-      } else if (cov_flag == "off"){
+      if (cov_flag == "off"){
         covMat_rbc = NA
         c_rbc_flag = FALSE
         se_rbc = rep(NA, length(y_grid))
-
+      } else {
+        covmat_rbc = cov_hat(x_data=x_data, y_data=y_data, x=x, y_grid=y_grid,
+                             p=p_RBC, q=q_RBC, mu=mu, nu=nu, h=bw,
+                             kernel_type=kernel_type, cov_flag=cov_flag)
+        covMat_rbc = covmat_rbc$cov
+        c_rbc_flag = covmat_rbc$singular_flag
+        se_rbc = sqrt(abs(diag(covMat_rbc)))*sd_y*mean(sd_x)
       }
     }
 
@@ -353,9 +348,10 @@ fhat = function(x_data, y_data, x, y_grid, p, q, mu, nu, h, kernel_type){
 #' @param mu Degree of derivative with respect to y.
 #' @param nu Degree of derivative with respect to x.
 #' @param kernel_type Kernel function choice.
+#' @param cov_flag Flag for covariance estimation option.
 #' @return Covariance matrix for all the grid points
 #' @keywords internal
-cov_hat = function(x_data, y_data, x, y_grid, p, q, mu, nu, h, kernel_type){
+cov_hat = function(x_data, y_data, x, y_grid, p, q, mu, nu, h, kernel_type, cov_flag){
   # setting constants
   n = length(y_data)
   d = ncol(x_data)
[…]
