+++ paper.bib
@@ -137,3 +137,14 @@ year={1995}
   journal={arXiv:1903.00954},
   year={2019}
 }
+
+@article{haldensify,
+doi = {10.21105/joss.04522},
+url = {https://doi.org/10.21105/joss.04522},
+year = {2022},
+publisher = {The Open Journal},
+volume = {7}, number = {77}, pages = {4522},
+author = {Nima S. Hejazi and Mark J. van der Laan and David Benkeser},
+title = {`haldensify`: Highly adaptive lasso conditional density estimation in `R`},
+journal = {Journal of Open Source Software}
+}
