+++ R/lpcde/man/lpcde.Rd
@@ -36,7 +36,9 @@ points will be chosen as 0.05-0.95 percentiles of the data, with a step size of 
 y-direction.}
 
 \item{x}{Numeric, specifies the grid of evaluation points in the x-direction. When set to default,
-the evaluation point will be chosen as the median of the x data.}
+the evaluation point will be chosen as the median of the x data. To generate
+estimates for multiple conditioning values, please loop over the x values and
+evaluate the lpcde function at each point.}
 
 \item{bw}{Numeric, specifies the bandwidth used for estimation. Can be (1) a positive
 scalar (common bandwidth for all grid points); or (2) a positive numeric vector/matrix
