### 2024-11-15 9b8c2e2
roxygen2 compatability
