+++ R/lpcde/man/plot.lpcde.Rd
@@ -40,22 +40,22 @@ they will be applied to each data series accordingly.}
 
 \item{lty}{Line type for point estimates, only effective if \code{type} is \code{"line"} or
 \code{"both"}. \code{1} for solid line, \code{2} for dashed line, \code{3} for dotted line.
-For other options, see the instructions for \code{\link{ggplot2}} . If
+For other options, see the instructions for \code{ggplot2} . If
 more than one is provided, they will be applied to each data series accordingly.}
 
 \item{lwd}{Line width for point estimates, only effective if \code{type} is \code{"line"} or
 \code{"both"}. Should be strictly positive. For other options, see the instructions for
-\code{\link{ggplot2}} . If more than one is provided, they will be applied
+\code{ggplot2} . If more than one is provided, they will be applied
 to each data series accordingly.}
 
 \item{lcol}{Line color for point estimates, only effective if \code{type} is \code{"line"} or
 \code{"both"}. \code{1} for black, \code{2} for red, \code{3} for green, \code{4} for blue.
-For other options, see the instructions for \code{\link{ggplot2}} . If
+For other options, see the instructions for \code{ggplot2} . If
 more than one is provided, they will be applied to each data series
 accordingly.}
 
 \item{pty}{Scatter plot type for point estimates, only effective if \code{type} is \code{"points"} or
-\code{"both"}. For options, see the instructions for \code{\link{ggplot2}} . If
+\code{"both"}. For options, see the instructions for \code{ggplot2} . If
 more than one is provided, they will be applied to each data series
 accordingly.}
 
@@ -66,7 +66,7 @@ accordingly.}
 \item{pcol}{Scatter plot color for point estimates, only effective if \code{type} is \code{"points"} or
 \code{"both"}. \code{1} for black, \code{2} for red, \code{3}
 for green, \code{4} for blue.
-For other options, see the instructions for \code{\link{ggplot2}} . If
+For other options, see the instructions for \code{ggplot2} . If
 more than one is provided, they will be applied to each data series
 accordingly.}
 
@@ -94,7 +94,7 @@ accordingly.
 
 \item{CIcol}{Color of the confidence region. \code{1} for black, \code{2} for red, \code{3}
 for green, \code{4} for blue.
-For other options, see the instructions for \code{\link{ggplot2}} . If
+For other options, see the instructions for \code{ggplot2} . If
 more than one is provided, they will be applied to each data series
 accordingly.}
 
