### 2024-11-15 nppackages/lpcde#5 @rajitachandak
We have added community guidelines on contributing and seeking support to the README of the repository.
