### 2024-11-15 nppackages/lpcde#6 @rajitachandak
We have fixed the license issue with the latest version of the package (0.1.5) on CRAN.
