### 2024-11-15 nppackages/lpcde#7 @rajitachandak
Thank you for bringing these points to our attention.
  (a) We have added DOIs to references where available.
  (b) We have fixed the reference.
  (c) We have fixed the spelling mistake.
  (d) We have added haldensify to the table and compared its capabilities along with the other packages.
