### 2024-11-15 nppackages/lpcde#8 @rajitachandak
(a) We have adopted the new syntax and ensured compatibility of our package with latest version of ggplot2.
(b) We have added small unit tests (where it is reasonable) to test properties of estimator.
