### 2024-11-15 nppackages/lpcde#10 @rajitachandak
Thank you for bringing this error to our attention. In our investigation of this issue, it seemed the error arose due to numerical instability (when there are no data points in a given neighborhood). We have resolved this issue and fixed the typo in the documentation.
