### 2024-11-15 nppackages/lpcde#9 @rajitachandak
We would like to address this comment in three separate points:

 (a) The difference in run time identified for two methods of implementing lpcde is due to the fact that the original implementation for lpcde also computes the full covariance matrix, which when given many grid points, can be quite slow to compute (of order n^3). To allow more flexibility (as in your example, where the covariance estimate is not needed), we have added a new optional input cov_flag to the lpcde function which can take on three different values:
        (1) "full": this is the original implementation, the function will compute the entire covariance matrix (and therefore allow confidence interval and band construction),
        (2) "diag": this will only compute the diagonal entries of the covariance matrix (i.e. the standard errors, only pointwise confidence intervals can be computed), and
        (3) "off": no entries of the covariance matrix are estimated.

  (b) For estimating the density over different conditioning values (x), a for loop will need to be used outside of calling the lpcde function. This is to ensure the inference capabilities (i.e. confidence intervals and uniform confidence bands) of the function can still be used, if desired. We have added a note regarding this in the description of the lpcde function.

  (c) It turns out our implementation lpcde already works for prediction on a new dataset directly. Below is a simple illustration of how this can be coded up:

Suppose we choose different data to "train" our estimator and to "test" it on an unseen dataset. In this case, a common method is to randomly subset the data into training and testing.
We start with the same simulation setup as in your example:
```
     library("lpcde")
      set.seed(32)
      n=1000
      x_data = matrix(rbeta(n, 2, 4))
      y_data = matrix(rgamma(n, 10, 1/x_data))
```
Now, we want to split the data. This can be done with the following three lines of code:

```
      #use 95% of dataset as training set and 5% as test set
      sample = sample(c(TRUE, FALSE), nrow(y_data), replace=TRUE, prob=c(0.95,0.05))
      y_train  = y_data[sample, ]
      x_train = x_data[sample, ]
      y_test   = y_data[!sample, ]
```
Now we simply input the y_test as the grid of evaluation points to lpcde:
```
   prediction_model = lpcde::lpcde(x_data=x_train, y_data=y_train, y_grid=y_test, x=0.5, bw=0.5, cov_flag="off")
```
If we wanted to extract only the point estimates of this model (as is usually the output of the predict function), we could run

```
coef(prediction_model)
```

Additionally, we note that the predict function only works effectively when there is a closed-form solution for evaluating the function at new data points. Since our estimator is nonparametric (of the local smoothing variety), it will require to construct a new estimate for each new evalulation point of interest (e.g., the bandwidth needs to be re-estimated, and the local data contributes differently as determined by the equivalent kernel). If the objective is to simply extract the point estimates, we recommend in the documentation that the coef method can be used to extract only the estimates from the lpcde object.
