### 2024-11-15 #7241 @rajitachandak
Dear @salbalkus,

Thank you for your thorough review and feedback of our package and paper. We have addressed each of the issues raised on our repository and provided a note regarding the changes we have made for each issue individually. Just to summarize our revision -- We made all the changes requested. We have added a discussion on the issue regarding implementing the predict method that we hope will help explain our implementation of this request.

Please feel free to reach out if you have any questions.
