### 2024-11-15 nppackages/lpcde#6 @salbalkus
Excellent! Note that it looks like you've included an MIT license, but the license shield in the first line of the README says GPLv2 and the "license" tab says GPL-3.0; you'll probably want to make these consistent with the MIT license.
