### 2024-11-15 nppackages/lpcde#9 @salbalkus
This is fantastic. Thank you for clearing up this issue; the example you provided is perfect for helping users understand how to get predictions, and the `cov_flag` parameter is a great solution to allow users to easily compute predictions quickly. Very clear explanation!
