### 2024-11-15 #7241 @salbalkus
Hi @rajitachandak,

Thank you for the in-depth explanations and attention to detail on this response. All of my previous concerns have been addressed! As noted in  [JOSS review: Fix license nppackages/lpcde#6](https://github.com/nppackages/lpcde/issues/6) the only remaining issue I see is to change the GPL license GitHub shield and license tabs in the README to be consistent with the MIT license that you have included. Once that is complete, I will recommend the software for publication!
