### 2024-11-16 #7241 @rajitachandak
Dear @salbalkus,

Thanks for pointing out that inconsistency! I have update all license files in our repository to GPL >=3.0. The changes will reflect on CRAN shortly.
