### 2024-11-16 nppackages/lpcde#6 @rajitachandak
Thanks for pointing that out! I've made the license GPL >=3.0 consistent across all files. This should be reflecting on CRAN soon.
