### 2024-11-16 #7241 @salbalkus
Hi @rajitachandak,

Thanks again for the update. It looks like I'm still seeing an MIT license in the README.md file located in the top level of the repository (the one that displays [at this link ](https://github.com/nppackages/lpcde) directly). It is located in the section titled "License", just after the "References" section that cites your papers.

Since the GPL license is included in a separate file elsewhere it may be best to just remove this section of the top level README.md.
