### 2025-01-28 #7241 @spholmes
Hi @sdbpimentel
I'd love to have your help in reviewing the lpcde package that was submitted to JOSS, could you take a quick look and advise me?
Many thanks
Susan
