### 2025-01-30 #7241 @spholmes
Hi @matteobonvini,
I wonder if you could help us out here at JOSS for this submission lpcde on local polynomial conditional density estimators.
Many thanks
Susan
