### 2025-02-20 #7241 @rajitachandak
Dear @spholmes,

Thank you for handling our paper. We completed the review process with the first reviewer in Nov 16, 2024, and we have been waiting for the review process to continue since then. Would it be possible to complete the review process? If more recommended reviewers are needed, could you please give us acceptable parameters for our recommendation?

Many thanks for your consideration.
