### 2025-02-23 #7241 @spholmes
Dear @soichiroy,
Professor K Imai suggested we contact you to see if you give us a hand finishing the review of the lpcde submission to JOSS.
It's a pretty straightforward process based on a checklist of things to check, much of the busywork is done automatically in Github,
please let us know if you can help us out here.
Thanks so much,
Susan
