### 2025-02-24 #7241 @soichiroy
> Dear [@soichiroy](https://github.com/soichiroy), Professor K Imai suggested we contact you to see if you give us a hand finishing the review of the lpcde submission to JOSS. It's a pretty straightforward process based on a checklist of things to check, much of the busywork is done automatically in Github, please let us know if you can help us out here. Thanks so much, Susan

Hi @spholmes I'm happy to review the submission. Please let me know what's the next step. Thank you!
