### 2025-02-24 #7241 @spholmes
Thanks @soichiroy..you can now generate a checklist using the command
"@editorialbot generate my checklist"
(without the quotes)
please see also the documentation about various points on the list (COI etc)
let me know if there is trouble or if I can help in any way, we are very grateful for your time,
Susan
