### 2025-02-24 nppackages/lpcde#11 @soichiroy
Organize replication related scripts into a subdirectory
Per JOSS submission: https://github.com/openjournals/joss-reviews/issues/7241

> Repository: Is the source code for this software available at the https://github.com/nppackages/lpcde?

This is a nit, so please feel free to ignore if there's a reason to put the replication related scripts under `R/`, but it would be clearer where to find the package related scripts (and for the replication related scripts), if they're bundled under `replication/` (or some directory name that indicates as such).
