### 2025-02-24 nppackages/lpcde#12 @soichiroy
Clarification of the package authors
Per JOSS submission: https://github.com/openjournals/joss-reviews/issues/7241

> Contribution and authorship: Has the submitting author (@rajitachandak) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

The package lists @rajitachandak as [the solo author and creator of the package](https://github.com/nppackages/lpcde/blob/main/R/lpcde/DESCRIPTION#L5-L8), while the readme file lists a reference that mention [other people](https://github.com/nppackages/lpcde/blob/main/R/lpcde/README.md?plain=1#L21). Is this intended?

Another question is that if [this article](https://github.com/openjournals/joss-reviews/issues/7241#issuecomment-2356819614) should be listed in the reference section of the README file of the package?
