### 2025-02-24 nppackages/lpcde#13 @soichiroy
Incorporate examples in the package readme file.
Per JOSS submission: https://github.com/openjournals/joss-reviews/issues/7241

> Functionality: Have the functional claims of the software been confirmed?
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Nit: I wonder if it would be worth expanding the package read me file a little bit to incorporate the working example and to explain the output. Currently, a user need to navigate the [sample script](https://github.com/nppackages/lpcde/blob/main/R/lpcde_illustration.R), [the package paper](https://arxiv.org/pdf/2204.10375), and perhaps the package manual / example. If they're in a single Readme or vignette, it would be easier as a user to understand the functionality of the package and how to interpret results.

For example, it would be great if, for example, what we should expect from the two models in the example is documented:

```r
model1 = lpcde::lpcde(x_data=x_data, y_data=y_data, y_grid=y_grid, x=0, bw=1)
summary(model1)
model_reg = lpcde::lpcde(
  x_data=x_data, y_data=y_data, y_grid=y_grid, x=0, bw=1, nonneg=TRUE, normalize=TRUE)
summary(model_reg)
```

Another nit is that if there's any reason why `y_grid` is set as `y_grid = seq(from=-2, to=2, length.out=10)` while `seq(-2, 2, by = 0.5)` will provide cleaner output in terms of the "Grid" column in the output table.
