### 2025-02-24 #7241 @soichiroy
@spholmes I completed the review. Most of the issues have been addressed by the previous iteration of the review.

@rajitachandak I added several minor comments on the package repo as issues:

- [ ] https://github.com/nppackages/lpcde/issues/11
- [ ] https://github.com/nppackages/lpcde/issues/12
- [ ] https://github.com/nppackages/lpcde/issues/13

Overall the package and the paper look solid to me. While it might be out of scope of the paper reviewer, I recommend doing a bit of cleaning of the code base. For example, I noticed a few scripts containing un-used code ([example](https://github.com/nppackages/lpcde/blob/main/R/lpcde/R/lpbwcde_fns.R#L372-L690), [example](https://github.com/nppackages/lpcde/blob/main/R/lpcde_illustration.R#L14)),  which in my opinion make it harder for someone to contribute in the future. Also applying something like [styler](https://github.com/r-lib/styler) might be a good option.
