### 2025-02-25 ee959f7
edits from JOSS review
