+++ README.md
@@ -6,7 +6,7 @@
 
 The `lpcde` package provides R implementation of bandwidth selection, point estimation and inference procedures for local polynomial conditional distribution and density methods.
 
-This work was supported by the National Science Foundation through grant [SES-1947805](https://www.nsf.gov/awardsearch/showAward?AWD_ID=1947805) and [SES-1947662](https://www.nsf.gov/awardsearch/showAward?AWD_ID=1947662), and by the National Institutes of Health through grant [R01 GM072611-16](https://reporter.nih.gov/project-details/10093056).
+This work was supported by the the National Science Foundation through grants SES-1947805, SES-1947662, DMS-2210561, and SES-2241575, and from the National Institute of Health (R01 GM072611-16).
 
 ## Website
 
