### 2025-02-25 nppackages/lpcde#13 @rajitachandak
Thank you for bringing this to our attention. We have expanded the package README to include some simple examples. We refer the reader to specific replication files and articles for a complete explanation of the functions.

We have also modified the replication scripts so that the `y_grid` object produces cleaner output.
