### 2025-02-25 nppackages/lpcde#12 @rajitachandak
Thank you for identifying this. Indeed the other authors should have been included in the package description file. I have fixed this now, it may take a few days to reflect on CRAN.

We have also included the JOSS article in the reference section. We will include the doi once available.
