### 2025-02-25 nppackages/lpcde#11 @rajitachandak
Thank you for this suggestion. Indeed this would be beneficial to users of the software. We have moved the replication scripts to a new folder at the top level of the repository.
