### 2025-02-25 #7241 @rajitachandak
Dear @soichiroy,

Thank you for your quick review of our package and paper. We have addressed each of the issues raised on our repository. Just to summarize our edits -- we have made all the changes requested.

Please let us know if there are any additional comments you would like us to address.

Best,
Rajita
