### 2025-02-25 nppackages/lpcde#11 @soichiroy
Thanks for updating. I think you might also want to update the corresponding link on the readme file: https://github.com/nppackages/lpcde/blob/cd47a22d03a596fd82e771d5d6b886ef4b8799b5/README.md?plain=1#L46
