### 2025-02-25 #7241 @soichiroy
> Dear [@soichiroy](https://github.com/soichiroy),
>
> Thank you for your quick review of our package and paper. We have addressed each of the issues raised on our repository. Just to summarize our edits -- we have made all the changes requested and have submitted the new version of the package to CRAN. It may take a few days for this new version to appear on CRAN as it does on GitHub.
>
> Please let us know if there are any additional comments you would like us to address.
>
> Best, Rajita

@rajitachandak Thank you for addressing the comments. @spholmes the author addressed all the major comments I provided (I suggested a fix on a minor typo in https://github.com/nppackages/lpcde/issues/11#issuecomment-2683491171).
