### 2025-03-04 #7241 @rajitachandak
Dear @spholmes,

We have checked all the author tasks. The version of the package for this publication is 0.1.6 and I've archived it with Zenodo (DOI: 10.5281/zenodo.14959119).

Please let me know if there's anything else you need from us.

Best,
Rajita
