### 2025-03-05 #7241 @crvernon
:wave: @rajitachandak - I just need you to address the following before I move to accept this for publication:

In your Zenodo archive:

- [ ] The author order of the names in your Zenodo archive do not match the order in your paper.  Could you please adjust these in your archive?  No new release needed.

In the paper:

- [ ] LINE 87:  the "r" as in the R programming language should be capitalized here.  You can do this in your bib file by wrapping characters you wish to maintain formatting of in curly brackets.

Let me know once the changes have been made and I'll accept. Thanks.
