### 2025-03-06 #7241 @rajitachandak
Hi @crvernon,

I've made the changes to the Zenodo archive and fixed the bib reference.

Thank you!

Best,
Rajita
